requirements.txt
# Core dependencies
aiofiles>=0.7.0
tabulate>=0.8.9
pyyaml>=6.0
yara-python>=4.2.0; platform_system != "Windows" or python_version < "3.11"  # Skip on Windows Python 3.11+
requests>=2.28.0
tqdm>=4.65.0  # For progress bars
colorama>=0.4.6  # For colored output